requirements.txt
google-cloud-firestore
google-api-python-client
rich